refactored abbreviated number method
