+++ src/utils/number.ts
@@ -26,6 +26,8 @@ export const getDecimalPlaces = (value: number) => {
   return match ? match[0].length : 0;
 };
 
+
+
 export const formatNumberInputStr = (text: string, precision = 10) => {
   if (text.includes(',')) {
     text = text.replace(',', '.');
@@ -47,5 +49,27 @@ export const formatNumberInputStr = (text: string, precision = 10) => {
     return '0';
   } else {
     return text;
+  }
+};
+
+
+
+
+export const getAbbreviatedNumber = (input: string | number) => {
+  const num = parseFloat(input); // Convert the string to a number
+
+  // Check if the input is not a valid number
+  if (Number.isNaN(num)) {
+    return input; // Return the original string if it's not a number
+  }
+
+  if (num >= 1_000_000) {
+    return `${(num / 1_000_000).toFixed(1).replace(/\.0$/, '')}M`; // Format in millions
+  } else if (num >= 100_000) {
+    return `${(num / 1_000_000).toFixed(1).replace(/\.0$/, '')}M`; // Convert 100K to 0.1M
+  } else if (num >= 1_000) {
+    return `${(num / 1_000).toFixed(1).replace(/\.0$/, '')}K`; // Format in thousands
+  } else {
+    return num.toString(); // Return smaller numbers as-is
   }
 };
